HTTP Mode
Added --http flag to get the HTTP method.
Added --debug flag (prints sent request).
Added HTTP Method Validity Checker, can be disabled using --force

When --http is used, the content data is used as the HTTP body and a
HTTP/1.1 request to / path is made with the specified method.
Currently, the address is not parsed correctly. It will be fixed in the
next commit.