fix: 499 status code from some servers
tURL used to shut the connection down just after making the request.
This was causing some servers to respond with 499 status code (that
doesn't even reach the client).

The solution was to remove the
`stream.shutdown(std::net::Shutdown::Write)?;` line.

However, this function was used to prevent the client from waiting the
server to close the connection after responding (some servers never
close the connection).

I had to add a timeout to the response reading. If the server doesn't
respond in time, the response that was read to that point is used.

When request results in timeout, it is not always a loss. Because the
printing/saving goes on with the response that has been read to that
point.

Timeout is set to 5 seconds by default. It can be changed with the
--timeout flag.

Note: This won't change anything on HTTP Mode because using "Connection:
close" header will make the server close the connection after
responding.