feat: indefinite timeout option & --hide-http-headers
Added indefinite timeout option. (set timeout to 0 with --force)

When timeout is set to 0, turl will warn the user and exit unless the
--force flag is set. When --force is set, turl will wait until the
server closes the connection.

Added --hide-http-headers flag.

When this flag is set, the response is splitted at every '\r'.
With a for loop it iterates until it finds a "\n".as_bytes() and after
that every line is saved as the body in a vector.
Then the response is overwritten with the body.